Fixed restart fails to prompt for new auth token
When restarting after a significant period of time (e.g. after waking
from being suspended for hours/days, etc.), we were mistakenly re-using
the cached authentication token in an attempt to automatically reconnect
to the VPN.

This would cause an authentication failure, because the cached token had
expired.

To fix this, we now wait for the VPN client to request our
authentication token before launching the browser SAML negotiation. This
ensures that when reconnecting, we will re-run the browser-based auth
flow, generating a new auth token.

To enable this, we've had to extract our auth flow in to a function, and
switch from using a BASH coprocess to just using a file descriptor and
command-substitution, which is actually a bit cleaner anyway.